Use a single future implementation
